split lightning_models.py into pieces
